build: Fix duplicate generated modules for cross-partition conflicts
The partition check for conflicts fails when we have cross-partition
conflicts, as the same-partition module may also be a conflict.

To fix the issue, keep a separate map of all candidate modules with the
same name and check all of them for duplicate entries, even if some fail
the partition check.